MP-110 :bug: Fix: 팁 게시글 조회수 필드 데이터 타입 변경
- 조회수가 많아질 경우 오버플로우를 방지하기 위해 조회수 필드의 데이터 타입을 Integer->Long으로 변경

+++ src/main/java/kr/modusplant/domains/tip/domain/model/TipPost.java
@@ -20,7 +20,7 @@ public class TipPost {
 
     private final Integer recommendationNumber;
 
-    private final Integer viewCount;
+    private final Long viewCount;
 
     private final String title;
 
@@ -34,7 +34,7 @@ public class TipPost {
         private UUID authMemberUuid;
         private UUID createMemberUuid;
         private Integer recommendationNumber;
-        private Integer viewCount;
+        private Long viewCount;
         private String title;
         private JsonNode content;
         private Boolean isDeleted;
